pyproject.toml
[project]
name = "fin-tracker-ui"
version = "0.1.0"
description = ""
authors = [{ name = "Dorian Beganovic" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "streamlit",
    "duckdb",
    "matplotlib",
    "skfolio",
    "riskfolio-lib>=6.3.1,<7",
    "seaborn>=0.13.2,<0.14",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "pymupdf>=1.26.5",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "camelot-py",
    "fear-and-greed>=0.4",
]

[dependency-groups]
dev = [
    "watchdog>=4.0.0,<5",
    "ruff",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
